Timer issue with forge
In its current form, `c.heroclass["cooldown"]` stores the result of (current time from `time.time()`) + `cooldown_time` which means the time stored in config is the expected time for cooldown to expire.
However during comparision in line 2643 `cooldown_time` is added to `c.heroclass["cooldown"]` which would be correct if `c.heroclass["cooldown"]` stored time of last command run. In this case it doubles up the time for cooldown.